Add range diffing utility for gitlab mr versions
Supplied with a token and an merge request url this script will let the
user choose two merge request versions do diff using range-diff. It will
create an empty repository in .mr-version-differ in which each requested
mr project will be added as a remote to be able to fetch the specified
version refs.